test: fix studio tests
